Criação do repository de Compromisso e validações para excluir lançamentos vinculados a compromissos

+++ ConFin.Domain/Lancamento/LancamentoService.cs
@@ -1,21 +1,24 @@
 ﻿using ConFin.Common.Domain;
 using ConFin.Common.Domain.Dto;
+using ConFin.Domain.Compromisso;
 using System;
 using System.Collections.Generic;
+using System.Linq;
 
 namespace ConFin.Domain.Lancamento
 {
-    public class LancamentoService: ILancamentoService
+    public class LancamentoService : ILancamentoService
     {
         private readonly ILancamentoRepository _lancamentoRepository;
+        private readonly ICompromissoRepository _compromissoRepository;
         private readonly Notification _notification;
 
-        public LancamentoService(Notification notification, ILancamentoRepository lancamentoRepository)
+        public LancamentoService(Notification notification, ILancamentoRepository lancamentoRepository, ICompromissoRepository compromissoRepository)
         {
             _notification = notification;
             _lancamentoRepository = lancamentoRepository;
+            _compromissoRepository = compromissoRepository;
         }
-
 
         public IEnumerable<LancamentoDto> GetAll(int idUsuario, byte? mes = null, short? ano = null, int? idConta = null, int? idCategoria = null)
         {
@@ -38,7 +41,7 @@ namespace ConFin.Domain.Lancamento
                 return;
             }
 
-            var periodo = _lancamentoRepository.GetPeriodo((byte) lancamento.IdPeriodo);
+            var periodo = _lancamentoRepository.GetPeriodo((byte)lancamento.IdPeriodo);
             if (periodo == null)
             {
                 _notification.Add("Os dados do periodo informado não foram encontrados");
@@ -47,7 +50,7 @@ namespace ConFin.Domain.Lancamento
 
             _lancamentoRepository.OpenTransaction();
 
-            var idCompromisso = _lancamentoRepository.PostCompromisso(new CompromissoDto
+            var idCompromisso = _compromissoRepository.Post(new CompromissoDto
             {
                 Descricao = lancamento.Descricao,
                 IdPeriodo = (byte)lancamento.IdPeriodo,
@@ -68,7 +71,7 @@ namespace ConFin.Domain.Lancamento
                 while (lancamento.Data <= dataLimite)
                 {
                     var idLancamento = _lancamentoRepository.Post(lancamento);
-                    _lancamentoRepository.PostCompromissoLancamento(idCompromisso, idLancamento, cont++);
+                    _compromissoRepository.PostCompromissoLancamento(idCompromisso, idLancamento, cont++);
                     lancamento.Data = AddDate(lancamento.Data, periodo);
                 }
 
@@ -87,7 +90,7 @@ namespace ConFin.Domain.Lancamento
             for (; cont <= lancamento.TotalParcelasOriginal; cont++)
             {
                 var idLancamento = _lancamentoRepository.Post(lancamento);
-                _lancamentoRepository.PostCompromissoLancamento(idCompromisso, idLancamento, cont);
+                _compromissoRepository.PostCompromissoLancamento(idCompromisso, idLancamento, cont);
                 lancamento.Data = AddDate(lancamento.Data, periodo);
             }
 
@@ -95,12 +98,48 @@ namespace ConFin.Domain.Lancamento
 
         }
 
-        public void Delete(int idLancamento)
+        public void Delete(int idLancamento, string indTipoDelete)
         {
+            var compromisso = _compromissoRepository.GetCompromissoLancamento(idLancamento);
+
+            if (compromisso == null)
+            {
+                _lancamentoRepository.Delete(idLancamento);
+                return;
+            }
+
+            if (string.IsNullOrEmpty(indTipoDelete) || !new List<string>() { "U", "P", "T" }.Contains(indTipoDelete))
+            {
+                _notification.Add("Falha ao excluir lançamento fixo/parcelado, o tipo de remoção não foi enviado ou esta inválido");
+                return;
+            }
+
+            var lancamentosVinculados = _compromissoRepository.GetCompromissoLancamentos(compromisso.Id).ToList();
+            List<CompromissoLancamentoDto> lancamentosVinculadosAux;
+
+            switch (indTipoDelete)
+            {
+                case "U": lancamentosVinculadosAux = lancamentosVinculados.Where(x => x.IdLancamento == idLancamento).ToList(); break;
+                case "P": lancamentosVinculadosAux = lancamentosVinculados.Where(x => x.IdLancamento >= idLancamento).ToList(); break;
+                default: lancamentosVinculadosAux = lancamentosVinculados.ToList(); break;
+            }
+
             _lancamentoRepository.OpenTransaction();
-            _lancamentoRepository.DeleteCompromissoLancamento(idLancamento);
-            _lancamentoRepository.Delete(idLancamento);
-            _lancamentoRepository.CommitTransaction();
+
+            foreach (var lancamento in lancamentosVinculadosAux)
+            {
+                _compromissoRepository.DeleteCompromissoLancamento(lancamento.IdLancamento);
+                _lancamentoRepository.Delete(lancamento.IdLancamento);
+                lancamentosVinculados.RemoveAll(x => x.IdLancamento == lancamento.IdLancamento);
+            }
+
+            if(!lancamentosVinculados.Any())
+                _compromissoRepository.Delete(compromisso.Id);
+
+            if (_notification.Any)
+                _lancamentoRepository.RollbackTransaction();
+            else
+                _lancamentoRepository.CommitTransaction();
         }
 
         private static DateTime AddDate(DateTime dataAtual, PeriodoDto periodo)
